feat: computers vers API :sparkles:
