+++ src/commons/components/Forms/ComputerAddForm/ComputerAddFormComponent.tsx
@@ -30,7 +30,7 @@ export const ComputerAddFormComponent = (props: ComputerAddFormComponentProps) =
                         props.setComputerSerial(e.detail.value!)
                         props.setNewComputerInfo({
                             ...props.newComputerInfo,
-                            serial: e.detail.value!
+                            serialNumber: e.detail.value!
                         })
                     }}
                     autoFocus
